+++ nebula_decoders/src/nebula_decoders_velodyne/velodyne_driver.cpp
@@ -50,7 +50,7 @@ Status VelodyneDriver::SetCalibrationConfiguration(
 }
 
 std::tuple<drivers::NebulaPointCloudPtr, double> VelodyneDriver::ParseCloudPacket(
-  const std::vector<uint8_t> & packet, int32_t packet_seconds)
+  const std::vector<uint8_t> & packet, double packet_seconds)
 {
   std::tuple<drivers::NebulaPointCloudPtr, double> pointcloud;
